Refactor: Extract custom Shaka buttons into separate files
