+++ Resources/Private/JavaScript/VideoPlayer/SachsenShakaPlayer.js
@@ -2,8 +2,14 @@ import $ from 'jquery';
 import BookmarkModal from './BookmarkModal';
 import Chapters from './Chapters';
 import SimpleModal from './SimpleModal';
-import { renderScreenshot } from './Screenshot';
-import { buildTimeString } from './util';
+
+import CaptureButton from './controls/CaptureButton';
+import PresentationTimeTracker from './controls/PresentationTimeTracker';
+import Replay10Button from './controls/Replay10Button';
+import SkipPreviousButton from './controls/SkipPreviousButton';
+import SkipNextButton from './controls/SkipNextButton';
+import Forward10Button from './controls/Forward10Button';
+import BookmarkButton from './controls/BookmarkButton';
 
 import '../../Less/VideoPlayer/VideoPlayer.less';
 import './controls.css';
@@ -333,297 +339,3 @@ function registerKeybindings() {
     }
   });
 }
-
-class CaptureButton extends shaka.ui.Element {
-  static KEY = 'capture';
-
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    // The actual button that will be displayed
-    this.button_ = document.createElement('button');
-    this.button_.className = 'material-icons-round';
-    this.button_.title = 'Screenshot';
-    this.button_.textContent = 'photo_camera';
-    this.parent.appendChild(this.button_);
-
-    // Listen for clicks on the button
-    this.eventManager.listen(this.button_, 'click', () => {
-      renderScreenshot(document.getElementById('video'));
-    });
-  }
-};
-
-CaptureButton.Factory = class {
-  create(rootElement, controls) {
-    return new CaptureButton(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  CaptureButton.KEY,
-  new CaptureButton.Factory()
-);
-
-class SkipNextButton extends shaka.ui.Element {
-  static KEY = 'skip_next';
-
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    // The actual button that will be displayed
-    this.button_ = document.createElement('button');
-    this.button_.className = 'material-icons-round';
-    this.button_.title = 'Einzelbild weiter';
-    this.button_.textContent = 'skip_next'; // independent of KEY
-    this.parent.appendChild(this.button_);
-
-    // Listen for clicks on the button
-    this.eventManager.listen(this.button_, 'click', () => {
-      this.controls.elSxndPlayer.vifa.seekForward(1);
-    });
-  }
-};
-
-SkipNextButton.Factory = class {
-  create(rootElement, controls) {
-    return new SkipNextButton(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  SkipNextButton.KEY,
-  new SkipNextButton.Factory()
-);
-
-class SkipPreviousButton extends shaka.ui.Element {
-  static KEY = 'skip_previous';
-
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    // The actual button that will be displayed
-    this.button_ = document.createElement('button');
-    this.button_.className = 'material-icons-round';
-    this.button_.title = 'Einzelbild zurück';
-    this.button_.textContent = 'skip_previous'; // independent of KEY
-    this.parent.appendChild(this.button_);
-
-    // Listen for clicks on the button
-    this.eventManager.listen(this.button_, 'click', () => {
-      this.controls.elSxndPlayer.vifa.seekBackward(1);
-    });
-  }
-};
-
-SkipPreviousButton.Factory = class {
-  create(rootElement, controls) {
-    return new SkipPreviousButton(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  SkipPreviousButton.KEY,
-  new SkipPreviousButton.Factory()
-);
-
-class Forward10Button extends shaka.ui.Element {
-  static KEY = 'forward_10';
-
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    // The actual button that will be displayed
-    this.button_ = document.createElement('button');
-    this.button_.className = 'material-icons-round';
-    this.button_.title = '10 Sekunden vor';
-    this.button_.textContent = 'forward_10'; // independent of KEY
-    this.parent.appendChild(this.button_);
-
-    // Listen for clicks on the button
-    this.eventManager.listen(this.button_, 'click', () => {
-      this.controls.elSxndPlayer.skipSeconds(+10);
-    });
-  }
-};
-
-Forward10Button.Factory = class {
-  create(rootElement, controls) {
-    return new Forward10Button(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  Forward10Button.KEY,
-  new Forward10Button.Factory()
-);
-
-class Replay10Button extends shaka.ui.Element {
-  static KEY = 'replay_10';
-
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    // The actual button that will be displayed
-    this.button_ = document.createElement('button');
-    this.button_.className = 'material-icons-round';
-    this.button_.title = '10 Sekunden zurück';
-    this.button_.textContent = 'replay_10'; // independent of KEY
-    this.parent.appendChild(this.button_);
-
-    // Listen for clicks on the button
-    this.eventManager.listen(this.button_, 'click', () => {
-      this.controls.elSxndPlayer.skipSeconds(-10);
-    });
-  }
-};
-
-Replay10Button.Factory = class {
-  create(rootElement, controls) {
-    return new Replay10Button(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  Replay10Button.KEY,
-  new Replay10Button.Factory()
-);
-
-class BookmarkButton extends shaka.ui.Element {
-  static KEY = 'bookmark';
-
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    // The actual button that will be displayed
-    this.button_ = document.createElement('button');
-    this.button_.className = 'material-icons-round';
-    this.button_.title = 'Bookmark';
-    this.button_.textContent = 'bookmark_border';
-    this.parent.appendChild(this.button_);
-
-    // Listen for clicks on the button
-    this.eventManager.listen(this.button_, 'click', () => this.controls.elSxndPlayer.showBookmarkUrl());
-  }
-};
-
-
-BookmarkButton.Factory = class {
-  create(rootElement, controls) {
-    return new BookmarkButton(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  BookmarkButton.KEY,
-  new BookmarkButton.Factory()
-);
-
-const TimeMode = {
-  CurrentTime: 0,
-  RemainingTime: 1,
-  CurrentFrame: 2,
-  COUNT: 3,
-};
-
-class PresentationTimeTracker extends shaka.ui.Element {
-  static KEY = 'time_and_duration_frame';
-
-  /**
-   * @param {!HTMLElement} parent
-   * @param {!shaka.ui.Controls} controls
-   */
-  constructor(parent, controls) {
-    super(parent, controls);
-
-    /** @type {!HTMLButtonElement} */
-    this.currentTime_ = document.createElement('button');
-    this.currentTime_.classList.add('shaka-current-time');
-    this.parent.appendChild(this.currentTime_);
-
-    this.state = {};
-
-    this.render({
-      totalSeconds: 0,
-      activeMode: TimeMode.CurrentTime,
-    });
-
-    this.eventManager.listen(this.currentTime_, 'click', () => {
-      // We toggle the time display here --> change mode on click --> values get updated in timeandseekrangeupdated event
-      // current time: HH:MM:SS:FF
-      // remaining time
-      // current frame
-      this.render({
-        activeMode: (this.state.activeMode + 1) % TimeMode.COUNT,
-      });
-    });
-
-    this.eventManager.listen(this.controls, 'timeandseekrangeupdated', () => {
-      this.render({
-        totalSeconds: this.controls.getDisplayTime(),
-      });
-    });
-  }
-
-  render(state) {
-    const newState = Object.assign({}, this.state, state);
-
-    const { totalSeconds, activeMode } = newState;
-    if (totalSeconds !== this.state.totalSeconds || activeMode !== this.state.activeMode) {
-      const elSxndPlayer = this.controls.elSxndPlayer;
-
-      const showHour = elSxndPlayer.video.duration >= 3600;
-
-      let text, title;
-
-      switch (activeMode) {
-        case TimeMode.CurrentTime:
-        default:
-          text = buildTimeString(totalSeconds, showHour);
-          if (elSxndPlayer.vifa) {
-            text += ':' + ("0" + (elSxndPlayer.vifa.get() % elSxndPlayer.fps)).slice(-2);
-          }
-          if (elSxndPlayer.video.duration) {
-            text += ' / ' + buildTimeString(elSxndPlayer.video.duration, showHour);
-          }
-          title = 'Aktuelle Laufzeit / Gesamtlaufzeit';
-          break;
-
-        case TimeMode.RemainingTime:
-          text = buildTimeString(elSxndPlayer.video.duration - totalSeconds, showHour);
-          title = 'Restlaufzeit';
-          break;
-
-        case TimeMode.CurrentFrame:
-          text = `${elSxndPlayer.vifa.get()}`;
-          title = 'Frame-Nummer';
-          break;
-      }
-
-      let currentChapter = elSxndPlayer.chapters.timeToChapter(totalSeconds);
-      if (currentChapter) {
-        text += ` – ${currentChapter.title}`;
-      }
-
-      this.currentTime_.textContent = text;
-      this.currentTime_.title = title;
-    }
-
-    this.state = newState;
-  }
-};
-
-
-/**
- * @implements {shaka.extern.IUIElement.Factory}
- */
-PresentationTimeTracker.Factory = class {
-  create(rootElement, controls) {
-    return new PresentationTimeTracker(rootElement, controls);
-  }
-};
-
-shaka.ui.Controls.registerElement(
-  PresentationTimeTracker.KEY,
-  new PresentationTimeTracker.Factory()
-);
